+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 All notable changes to this project will be documented in this file.
 
+## [3.1.3](https://github.com/apify/apify-client-python/releases/tag/v3.1.3) (2026-08-18)
+
+### 🐛 Bug Fixes
+
+- Require value in EnvVarRequest and allow null error messages in request dicts ([#1001](https://github.com/apify/apify-client-python/pull/1001)) ([4540199](https://github.com/apify/apify-client-python/commit/45401992edd31f2f4273eed5549fd9615226f17b)) by [@apify-service-account](https://github.com/apify-service-account)
+- Warn when a requested timeout is capped at timeout_max ([#962](https://github.com/apify/apify-client-python/pull/962)) ([4281530](https://github.com/apify/apify-client-python/commit/4281530f0637215d00c2cb186c7917a5742e0bb8)) by [@vdusek](https://github.com/vdusek)
+- Do not mask unrelated import errors when guarding optional dependencies ([#1009](https://github.com/apify/apify-client-python/pull/1009)) ([f57e9e7](https://github.com/apify/apify-client-python/commit/f57e9e7338503d700f0b675ae7097a369e65734f)) by [@vdusek](https://github.com/vdusek)
+- Keep HttpResponse isinstance checks from consuming streamed responses ([#1010](https://github.com/apify/apify-client-python/pull/1010)) ([8a2bb7b](https://github.com/apify/apify-client-python/commit/8a2bb7bfb9e0078e1d1633e567760bb431ef60b4)) by [@vdusek](https://github.com/vdusek)
+- Fail fast on transport errors a retry cannot fix ([#1019](https://github.com/apify/apify-client-python/pull/1019)) ([ab63eff](https://github.com/apify/apify-client-python/commit/ab63effe9da2ee0ef60729c10ee3b132d85cb60b)) by [@vdusek](https://github.com/vdusek)
+- Classify and retry failures while reading a streamed error body ([#1020](https://github.com/apify/apify-client-python/pull/1020)) ([2760aa6](https://github.com/apify/apify-client-python/commit/2760aa67c7b2acc61cc69996f412120fabccc3d4)) by [@vdusek](https://github.com/vdusek)
+
+
 ## [3.1.2](https://github.com/apify/apify-client-python/releases/tag/v3.1.2) (2026-08-10)
 
 ### 🐛 Bug Fixes